Cleaned up warnings, updated bzflag.vcproj, and updated src/bzflag/Makefile.am.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/branches/gsoc_server_listing@18335 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDuiServerInfo.cxx
@@ -103,7 +103,7 @@ void HUDuiServerInfo::resize()
 
   float x = getX() + fontHeight;
   fs.setMin(10, 10);
-  for (int i=0; i<readouts.size(); i++) {
+  for (int i=0; i<(int) readouts.size(); i++) {
     if ((i + 1) % 7 == 1) {
       x = (0.125f + 0.25f * (float)(i / 7)) * (float)getWidth();
       y = y0;
@@ -345,7 +345,7 @@ void HUDuiServerInfo::doRender()
 
   fillReadouts();
 
-  for (int i=0; i<readouts.size(); i++)
+  for (int i=0; i<(int) readouts.size(); i++)
   {
     readouts[i]->render();
   }
